pyproject.toml
[tool.poetry]
name = "open-heavens-api"
version = "0.1.0"
description = "Open Heavens devotional API"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-dotenv = "^1.0.0"
psycopg = "^3.2"
pydantic = "^1.10.9"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
httpx = "^0.24.0"
alembic = "^1.11.1"
sqlalchemy = "^2.0.20"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
